Move `field-conditions/Omitter.js` to `publish/Values.js`.
